refactor(planning_data_analyzer): reduce EPDMS debug helper duplication
