+++ planning/autoware_planning_data_analyzer/CMakeLists.txt
@@ -14,6 +14,7 @@ ament_auto_add_library(${PROJECT_NAME} SHARED
   src/metrics/geometry/lanelet_geometry.cpp
   src/metrics/geometry/lanelet_queries.cpp
   src/metrics/geometry/object_tracks.cpp
+  src/metrics/geometry/trajectory_utils.cpp
   src/metrics/epdms/aggregation/epdms_aggregation.cpp
   src/metrics/epdms/debug/epdms_debug_writer.cpp
   src/metrics/epdms/subscores/ego_progress.cpp
