+++ planning/autoware_planning_data_analyzer/src/open_loop_evaluator.cpp
@@ -18,12 +18,13 @@
 #include "metrics/epdms/debug/epdms_debug_writer.hpp"
 #include "metrics/geometry/metric_utils.hpp"
 #include "metrics/geometry/object_tracks.hpp"
+#include "metrics/geometry/trajectory_utils.hpp"
 #include "metrics/trajectory_metrics.hpp"
 
 #include <autoware/motion_utils/trajectory/conversion.hpp>
 #include <autoware/motion_utils/trajectory/trajectory.hpp>
-#include <autoware_utils_geometry/geometry.hpp>
 #include <autoware_utils_math/normalization.hpp>
+#include <rclcpp/duration.hpp>
 #include <rclcpp/serialization.hpp>
 #include <rosbag2_cpp/reader.hpp>
 
@@ -145,78 +146,6 @@ std::shared_ptr<SynchronizedData> clone_with_trajectory(
   cloned->trajectory = std::make_shared<autoware_planning_msgs::msg::Trajectory>(trajectory);
   cloned->candidate_trajectories.reset();
   return cloned;
-}
-
-double point_time_s(const autoware_planning_msgs::msg::TrajectoryPoint & point)
-{
-  return rclcpp::Duration(point.time_from_start).seconds();
-}
-
-double lerp(const double lhs, const double rhs, const double ratio)
-{
-  return lhs + (rhs - lhs) * ratio;
-}
-
-autoware_planning_msgs::msg::TrajectoryPoint interpolate_trajectory_point(
-  const autoware_planning_msgs::msg::TrajectoryPoint & previous,
-  const autoware_planning_msgs::msg::TrajectoryPoint & next, const double horizon_s)
-{
-  const double previous_t = point_time_s(previous);
-  const double next_t = point_time_s(next);
-  const double ratio = next_t > previous_t
-                         ? std::clamp((horizon_s - previous_t) / (next_t - previous_t), 0.0, 1.0)
-                         : 0.0;
-
-  auto interpolated = previous;
-  interpolated.pose =
-    autoware_utils_geometry::calc_interpolated_pose(previous.pose, next.pose, ratio);
-  interpolated.longitudinal_velocity_mps =
-    lerp(previous.longitudinal_velocity_mps, next.longitudinal_velocity_mps, ratio);
-  interpolated.lateral_velocity_mps =
-    lerp(previous.lateral_velocity_mps, next.lateral_velocity_mps, ratio);
-  interpolated.acceleration_mps2 = lerp(previous.acceleration_mps2, next.acceleration_mps2, ratio);
-  interpolated.heading_rate_rps = lerp(previous.heading_rate_rps, next.heading_rate_rps, ratio);
-  interpolated.front_wheel_angle_rad =
-    lerp(previous.front_wheel_angle_rad, next.front_wheel_angle_rad, ratio);
-  interpolated.rear_wheel_angle_rad =
-    lerp(previous.rear_wheel_angle_rad, next.rear_wheel_angle_rad, ratio);
-  interpolated.time_from_start = rclcpp::Duration::from_seconds(horizon_s);
-  return interpolated;
-}
-
-autoware_planning_msgs::msg::Trajectory truncate_trajectory_by_horizon(
-  const autoware_planning_msgs::msg::Trajectory & trajectory, const double horizon_s)
-{
-  constexpr double kTimeEpsilon = 1.0e-6;
-  if (horizon_s <= 0.0 || trajectory.points.empty()) {
-    return trajectory;
-  }
-
-  autoware_planning_msgs::msg::Trajectory truncated;
-  truncated.header = trajectory.header;
-  truncated.points.reserve(trajectory.points.size());
-
-  for (const auto & point : trajectory.points) {
-    const double t = point_time_s(point);
-    if (t <= horizon_s + kTimeEpsilon) {
-      truncated.points.push_back(point);
-      continue;
-    }
-
-    if (!truncated.points.empty()) {
-      const double previous_t = point_time_s(truncated.points.back());
-      if (previous_t < horizon_s - kTimeEpsilon) {
-        truncated.points.push_back(
-          interpolate_trajectory_point(truncated.points.back(), point, horizon_s));
-      }
-    }
-    if (truncated.points.empty()) {
-      truncated.points.push_back(point);
-    }
-    return truncated;
-  }
-
-  return truncated;
 }
 
 metrics::EpdmsMetricSnapshot build_epdms_snapshot(const OpenLoopTrajectoryMetrics & metrics)
@@ -771,7 +700,7 @@ std::vector<OpenLoopEvaluator::EvaluationData> OpenLoopEvaluator::prepare_evalua
     }
 
     const auto truncated_trajectory =
-      truncate_trajectory_by_horizon(*data->trajectory, trajectory_evaluation_horizon_s_);
+      metrics::truncate_trajectory_by_horizon(*data->trajectory, trajectory_evaluation_horizon_s_);
     const auto truncated_data = clone_with_trajectory(data, truncated_trajectory);
 
     const size_t num_pts = truncated_data->trajectory->points.size();
